FIX : CPPCHECK warnings and perfomance notifications fixed partialy
svn path=/trunk/; revision=640

+++ ZEToolSDK/ZETFontFile/ZETFontFile.cpp
@@ -66,7 +66,7 @@ bool ZEFont::ReadFromFile(const char* Filename)
 	memcpy(Characters, FileHeader.Characters, sizeof(ZEFontCharacter) * ZEFONTFILE_CHARACTERCOUNT);
 
 	Textures.SetCount(FileHeader.TextureCount);
-	for (int I = 0; I < FileHeader.TextureCount; I++)
+	for (size_t I = 0; I < FileHeader.TextureCount; I++)
 	{
 		Textures[I].ReadFromResourceFile(File);
 	}
@@ -92,7 +92,7 @@ bool ZEFont::WriteToFile(const char* Filename)
 	fwrite(&FileHeader, sizeof(ZEFontFileHeader), 1, File);
 
 
-	for (int I = 0; I < FileHeader.TextureCount; I++)
+	for (size_t I = 0; I < FileHeader.TextureCount; I++)
 		Textures[I].WriteToResourceFile(File);
 
 	return true;
